requirements_updated.txt
absl-py==2.3.1
astunparse==1.6.3
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
google-ai-generativelanguage==0.4.0
google-api-core==2.28.1
google-auth==2.47.0
google-genai==1.63.0
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.2.0
keras==3.10.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0

numpy<2.0

opencv-python==4.8.1.78
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.2.3
pillow==12.1.0
proto-plus==1.27.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.0.7
requests==2.32.5
rich==14.2.0
rsa==4.9.1
scikit-learn==1.2.2
scipy==1.15.3
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
threadpoolctl==3.6.0
tqdm==4.67.1
twilio
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.4
wrapt==2.0.1